event processor refactor
